pyproject.toml
[project]
name = "portfolio-binance-agno"
version = "0.1.0"
description = "Portfolio inteligente para Binance com análise via agentes IA (Agno)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.4",
    "python-binance>=1.0.19",
    "redis>=5.0.0",
    "agno>=0.1.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.1",
    "cryptography>=43.0.0",
    "pyyaml>=6.0.2",
    "sentence-transformers>=3.0.0",
    "qdrant-client>=1.11.0",
    "httpx>=0.27.0",
]
